+++ app/Http/Controllers/GastosIndirectosController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\GastosIndirectos;
+use App\Models\GastosIndirectosFuentes;
 use Illuminate\Http\Request;
 
 class GastosIndirectosController extends Controller
@@ -37,8 +38,8 @@ class GastosIndirectosController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'clave' => 'required',
+        $valido= $request->validate([
+            'clave' => 'required|unique:gastos_indirectos,clave',
             'nombre' => 'required',            
         ]);
         GastosIndirectos::create([
@@ -46,7 +47,12 @@ class GastosIndirectosController extends Controller
             'nombre' => $request->nombre  
         ]);
 
-        return redirect()->route('gastosIndirectos.index');
+        if($valido==false){
+            return redirect()->route('gastosIndirectos.index')->withInput();
+        }else{
+            return redirect()->route('gastosIndirectos.index');
+        }
+        
     }
     /**
      * Display the specified resource.
@@ -94,7 +100,12 @@ class GastosIndirectosController extends Controller
      */
     public function destroy(GastosIndirectos $gastosIndirecto)
     {
-        $gastosIndirecto->delete();
-        return redirect()->route('gastosIndirectos.index')->with('eliminar','ok');
+        $existeEnFuentesGastos= GastosIndirectosFuentes::where('indirectos_id',$gastosIndirecto->id_indirectos)->exists();
+        if($existeEnFuentesGastos == null){ //si no hay existe
+            $gastosIndirecto->delete();
+             return redirect()->route('gastosIndirectos.index')->with('eliminar','ok');
+        }else{
+             return redirect()->route('gastosIndirectos.index')->with('eliminar','error');
+        }
     }
 }
